Update ET_Client.php
Although support had been added for multiple objects in the ET_Post object, the corresponding post() method that provides the actual Properties for the SOAP request made it impossible to actually use.  A methods "buildOverrideProps" in DataExtesnion_Row populates the properties as an array if the props are constructed as such.  Also the ET_Post functionality that supported arrays was abstracted into ET_Constructor (buildObjects) so it can be shared with ET_Patch and ET_Delete.  Now post, delete and patch calls all support modifying multiple rows.